add: -Wall -Wextra -Werror to target_compile_options, fix int-size_t comparison

+++ src/DeepThought.hpp
@@ -25,7 +25,7 @@ class DeepThought {
 
     const variant<string, DeepThoughtError>
     _getQuestion(const string &input) {
-        for (int idx = 0; idx < input.size(); ++idx) {
+        for (size_t idx = 0; idx < input.size(); ++idx) {
 
             if (input[idx] == '"') {
                 return variant<string, DeepThoughtError>(DeepThoughtError(DeepThoughtError::Type::QUESTION_IN_QUOTES));
